requirements.txt
Django==1.9.2
django-cors-headers==2.1.0
django-filter==0.15.3
djangorestframework==3.6.3
pytz==2017.2
